server/api/requirements.txt
python-dotenv~=1.1.1
pydantic-settings~=2.11.0
SQLAlchemy~=2.0.44
dnspython~=2.8.0
idna~=3.11
httpcore~=1.0.9
httpx~=0.28.1
anyio~=4.11.0
h11~=0.16.0
attrs~=25.4.0
rich~=14.2.0
typing_extensions~=4.15.0
Pygments~=2.19.2
markdown-it-py~=4.0.0
PyYAML~=6.0.3
requests~=2.32.5
multidict~=6.7.0
propcache~=0.4.1
sniffio~=1.3.1
click~=8.3.0
mdurl~=0.1.2
typer~=0.19.2
shellingham~=1.5.4
MarkupSafe~=3.0.3
yarl~=1.22.0
aiohttp~=3.13.1
aiosignal~=1.4.0
frozenlist~=1.8.0
python-multipart~=0.0.20
aiohappyeyeballs~=2.6.1
asyncpg~=0.30.0
certifi~=2025.10.5
fastapi~=0.119.0
pydantic~=2.12.2
pydantic_core~=2.41.4
starlette~=0.48.0
email-validator~=2.3.0
websockets~=15.0.1
docopt~=0.6.2
yarg~=0.1.10
rignore~=0.7.1
urllib3~=2.5.0
uvicorn~=0.37.0
httptools~=0.7.1
watchfiles~=1.1.1
greenlet~=3.2.4
annotated-types~=0.7.0
charset-normalizer~=3.4.4
Jinja2~=3.1.6
sentry-sdk~=2.42.0